bin/stuff-database/requirements.txt
pynacl==1.4.0
requests==2.23.0